requirements.txt
nicegui
pydantic
pyyaml
